refactor: extract Horizon.Server config into separate config modules
